moved BlockChainSource code into sub-modules for easier development / review
